% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{.datom_build_s3_uri}
\alias{.datom_build_s3_uri}
\title{Build Full S3 URI}
\usage{
.datom_build_s3_uri(bucket, key)
}
\arguments{
\item{bucket}{S3 bucket name.}

\item{key}{S3 object key (from \code{.datom_build_storage_key()}).}
}
\value{
Character string S3 URI.
}
\description{
Convenience function that combines bucket and key into an S3 URI.
}
\keyword{internal}
