% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sha.R
\name{.datom_compute_file_sha}
\alias{.datom_compute_file_sha}
\title{Compute SHA-256 of File}
\usage{
.datom_compute_file_sha(path)
}
\arguments{
\item{path}{Path to file.}
}
\value{
Character SHA-256 hash.
}
\description{
Compute SHA-256 of File
}
\keyword{internal}
