% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{.datom_create_ref}
\alias{.datom_create_ref}
\title{Create Initial ref.json Content}
\usage{
.datom_create_ref(data_store)
}
\arguments{
\item{data_store}{A \code{datom_store_s3} component (the data portion of the store).}
}
\value{
A list suitable for JSON serialization.
}
\description{
Builds the initial \code{ref.json} structure from the data store component.
No \code{previous} entries on first creation.
}
\keyword{internal}
