% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-edit.R
\name{.datom_edit_log_fields}
\alias{.datom_edit_log_fields}
\title{The Columns of the Edit Log, in One Place}
\usage{
.datom_edit_log_fields()
}
\value{
A character vector of column names.
}
\description{
The Columns of the Edit Log, in One Place
}
\keyword{internal}
