% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decommission.R
\name{.datom_gh_available}
\alias{.datom_gh_available}
\title{Check Whether the gh CLI Is Available}
\usage{
.datom_gh_available()
}
\value{
Logical scalar.
}
\description{
Check Whether the gh CLI Is Available
}
\keyword{internal}
