% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gov.R
\name{.datom_gov_push}
\alias{.datom_gov_push}
\title{Push the Gov Clone to Remote}
\usage{
.datom_gov_push(conn)
}
\arguments{
\item{conn}{A \code{datom_conn} with \code{gov_local_path} set.}
}
\value{
Invisible TRUE.
}
\description{
Pushes the current branch of the gov clone to its remote.
}
\keyword{internal}
