% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-local.R
\name{.datom_local_delete}
\alias{.datom_local_delete}
\title{Delete a File from Local Storage}
\usage{
.datom_local_delete(conn, key)
}
\arguments{
\item{conn}{A \code{datom_conn} object with \code{backend = "local"}.}

\item{key}{Relative storage key (after \verb{prefix/datom/}).}
}
\value{
Invisible \code{TRUE} on success.
}
\description{
Delete a File from Local Storage
}
\keyword{internal}
