% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-local.R
\name{.datom_local_exists}
\alias{.datom_local_exists}
\title{Check if Local Storage Object Exists}
\usage{
.datom_local_exists(conn, key)
}
\arguments{
\item{conn}{A \code{datom_conn} object with \code{backend = "local"}.}

\item{key}{Relative storage key (after \verb{prefix/datom/}).}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Check if Local Storage Object Exists
}
\keyword{internal}
