% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-local.R
\name{.datom_local_list_objects}
\alias{.datom_local_list_objects}
\title{List Objects in Local Storage}
\usage{
.datom_local_list_objects(conn, prefix)
}
\arguments{
\item{conn}{A \code{datom_conn} object with \code{backend = "local"}.}

\item{prefix}{Relative prefix to list under.}
}
\value{
Character vector of relative keys (relative to \code{conn$root}).
}
\description{
Lists files under a given prefix in the store.
}
\keyword{internal}
