% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-local.R
\name{.datom_local_write_json}
\alias{.datom_local_write_json}
\title{Write an R List to Local Storage as JSON}
\usage{
.datom_local_write_json(conn, key, data)
}
\arguments{
\item{conn}{A \code{datom_conn} object with \code{backend = "local"}.}

\item{key}{Relative storage key (after \verb{prefix/datom/}).}

\item{data}{An R list to serialize to JSON.}
}
\value{
Invisible \code{TRUE} on success.
}
\description{
Serializes \code{data} to JSON and writes to the store directory. Creates parent
directories if needed.
}
\keyword{internal}
