% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-storage.R
\name{.datom_storage_download}
\alias{.datom_storage_download}
\title{Download File from Storage}
\usage{
.datom_storage_download(conn, key, local_path)
}
\arguments{
\item{conn}{A \code{datom_conn} object.}

\item{key}{Relative storage key (after \verb{prefix/datom/}).}

\item{local_path}{Local file path (destination).}
}
\value{
Invisible \code{TRUE} on success.
}
\description{
Download File from Storage
}
\keyword{internal}
