% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-storage.R
\name{.datom_storage_exists}
\alias{.datom_storage_exists}
\title{Check if Storage Object Exists}
\usage{
.datom_storage_exists(conn, key)
}
\arguments{
\item{conn}{A \code{datom_conn} object.}

\item{key}{Relative storage key (after \verb{prefix/datom/}).}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Check if Storage Object Exists
}
\keyword{internal}
