% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-storage.R
\name{.datom_storage_read_json}
\alias{.datom_storage_read_json}
\title{Read and Parse JSON from Storage}
\usage{
.datom_storage_read_json(conn, key)
}
\arguments{
\item{conn}{A \code{datom_conn} object.}

\item{key}{Relative storage key (after \verb{prefix/datom/}).}
}
\value{
Parsed R list.
}
\description{
Read and Parse JSON from Storage
}
\keyword{internal}
