% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-storage.R
\name{.datom_storage_upload}
\alias{.datom_storage_upload}
\title{Upload File to Storage}
\usage{
.datom_storage_upload(conn, local_path, key)
}
\arguments{
\item{conn}{A \code{datom_conn} object.}

\item{local_path}{Local file path to upload.}

\item{key}{Relative storage key (after \verb{prefix/datom/}).}
}
\value{
Invisible \code{TRUE} on success.
}
\description{
Upload File to Storage
}
\keyword{internal}
