% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-storage.R
\name{.datom_storage_write_json}
\alias{.datom_storage_write_json}
\title{Write an R List to Storage as JSON}
\usage{
.datom_storage_write_json(conn, key, data)
}
\arguments{
\item{conn}{A \code{datom_conn} object.}

\item{key}{Relative storage key (after \verb{prefix/datom/}).}

\item{data}{An R list to serialize to JSON.}
}
\value{
Invisible \code{TRUE} on success.
}
\description{
Write an R List to Storage as JSON
}
\keyword{internal}
