% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.datom_store_backend}
\alias{.datom_store_backend}
\title{Get Backend Type from Store Component}
\usage{
.datom_store_backend(component)
}
\arguments{
\item{component}{A store component object.}
}
\value{
\code{"s3"} or \code{"local"}.
}
\description{
Get Backend Type from Store Component
}
\keyword{internal}
