% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.datom_store_region}
\alias{.datom_store_region}
\title{Get Region from Store Component}
\usage{
.datom_store_region(component)
}
\arguments{
\item{component}{A store component object.}
}
\value{
Region string or NULL.
}
\description{
Returns the AWS region for S3, NULL for local.
}
\keyword{internal}
