% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.datom_store_root}
\alias{.datom_store_root}
\title{Get Root from Store Component}
\usage{
.datom_store_root(component)
}
\arguments{
\item{component}{A store component object.}
}
\value{
Root string.
}
\description{
Returns the storage root: bucket name for S3, directory path for local.
}
\keyword{internal}
