% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.is_datom_store_component}
\alias{.is_datom_store_component}
\title{Check if Object is a Store Component}
\usage{
.is_datom_store_component(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE or FALSE.
}
\description{
Returns TRUE for any datom store component type (datom_store_s3, future
datom_store_local, etc.).
}
\keyword{internal}
