% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-git.R
\name{.tbit_check_git2r}
\alias{.tbit_check_git2r}
\title{Check git2r Availability}
\usage{
.tbit_check_git2r()
}
\value{
Invisible TRUE if available.
}
\description{
Aborts with a helpful message if git2r is not installed.
}
\keyword{internal}
