% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sha.R
\name{.tbit_sync_metadata}
\alias{.tbit_sync_metadata}
\title{Sync Single Table Metadata to S3}
\usage{
.tbit_sync_metadata(conn, name)
}
\arguments{
\item{conn}{Connection object.}

\item{name}{Table name.}
}
\value{
Summary of sync operation.
}
\description{
Sync Single Table Metadata to S3
}
\keyword{internal}
