% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{is_datom_conn}
\alias{is_datom_conn}
\title{Check if Object is a datom Connection}
\usage{
is_datom_conn(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE or FALSE.
}
\description{
Check if Object is a datom Connection
}
\keyword{internal}
