% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{is_datom_store_s3_creds}
\alias{is_datom_store_s3_creds}
\title{Check if Object is a Credentials-Only S3 Store Component}
\usage{
is_datom_store_s3_creds(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE or FALSE.
}
\description{
Check if Object is a Credentials-Only S3 Store Component
}
\examples{
creds <- datom_store_s3_creds(
  access_key = "AKIAIOSFODNN7EXAMPLE",
  secret_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
)
is_datom_store_s3_creds(creds)
is_datom_store_s3_creds("not a store")
}
