% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{is_tbit_conn}
\alias{is_tbit_conn}
\title{Check if Object is a tbit Connection}
\usage{
is_tbit_conn(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE or FALSE.
}
\description{
Check if Object is a tbit Connection
}
\keyword{internal}
