% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_valid_tbit_repo}
\alias{is_valid_tbit_repo}
\title{Check if Path is a Valid tbit Repository}
\usage{
is_valid_tbit_repo(
  path,
  checks = c("all", "git", "tbit", "renv"),
  verbose = FALSE
)
}
\arguments{
\item{path}{Path to evaluate.}

\item{checks}{Which checks to perform. Any combination of "all", "git",
"tbit", "renv".}

\item{verbose}{If TRUE, prints which tests passed/failed.}
}
\value{
TRUE or FALSE.
}
\description{
Validates tbit repository structure. Used internally and by dpbuild.
}
