% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{print.datom_link}
\alias{print.datom_link}
\title{Print a Member Link}
\usage{
\method{print}{datom_link}(x, ...)
}
\arguments{
\item{x}{A \code{datom_link} from a member of a set read with \code{\link[=datom_get_set]{datom_get_set()}}.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Print a Member Link
}
\examples{
# See datom_get_set() for a runnable set example; a link is one of its
# members' `$fetch` elements.
print(names(formals(datom_get_set)))
}
