% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{print.datom_store_local}
\alias{print.datom_store_local}
\title{Print a Local Store Component}
\usage{
\method{print}{datom_store_local}(x, ...)
}
\arguments{
\item{x}{A \code{datom_store_local} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Displays store configuration.
}
\examples{
tmp <- tempfile("datom_store_")
store <- datom_store_local(path = tmp, validate = TRUE)
print(store)
unlink(tmp, recursive = TRUE)
}
