% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbit-package.R
\docType{package}
\name{tbit-package}
\alias{tbit}
\alias{tbit-package}
\title{tbit: Version-Controlled Data Management for Reproducible Workflows}
\description{
Provides version-controlled data management by abstracting tables as code in git while storing actual data in cloud storage. Enables setting up cloud-based repositories, syncing data with automatic versioning, tracking complete data lineage, and accessing any historical version for reproducibility. Designed for clinical and scientific workflows where reproducibility is paramount.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amashadihossein/tbit}
  \item Report bugs at \url{https://github.com/amashadihossein/tbit/issues}
}

}
\author{
\strong{Maintainer}: Afshin Mashadi-Hossein \email{amashadihossein@gmail.com}

}
\keyword{internal}
