% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{tbit_repository_check}
\alias{tbit_repository_check}
\title{Check tbit Repository Structure}
\usage{
tbit_repository_check(path)
}
\arguments{
\item{path}{Path to evaluate.}
}
\value{
List of TRUE/FALSE per check.
}
\description{
Returns detailed check results for each component.
}
\keyword{internal}
