% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.datom_github_username}
\alias{.datom_github_username}
\title{Get GitHub Username from PAT}
\usage{
.datom_github_username(pat)
}
\arguments{
\item{pat}{GitHub personal access token.}
}
\value{
Username string.
}
\description{
Calls \code{GET /user} to get the authenticated user's login.
}
\keyword{internal}
