% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.datom_mask_secret}
\alias{.datom_mask_secret}
\title{Mask a Secret for Display}
\usage{
.datom_mask_secret(secret)
}
\arguments{
\item{secret}{A string.}
}
\value{
Masked string.
}
\description{
Shows first 4 characters followed by \verb{****}.
}
\keyword{internal}
