% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.datom_validate_github_pat}
\alias{.datom_validate_github_pat}
\title{Validate GitHub PAT}
\usage{
.datom_validate_github_pat(pat)
}
\arguments{
\item{pat}{GitHub personal access token.}
}
\value{
A list with \code{login} and \code{id}.
}
\description{
Calls GitHub \code{GET /user} to verify the PAT is valid.
}
\keyword{internal}
