% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{is_datom_store}
\alias{is_datom_store}
\title{Check if Object is a datom Store}
\usage{
is_datom_store(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE or FALSE.
}
\description{
Check if Object is a datom Store
}
