% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{is_datom_store_local}
\alias{is_datom_store_local}
\title{Check if Object is a Local Store Component}
\usage{
is_datom_store_local(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE or FALSE.
}
\description{
Check if Object is a Local Store Component
}
