% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{is_datom_store_s3}
\alias{is_datom_store_s3}
\title{Check if Object is an S3 Store Component}
\usage{
is_datom_store_s3(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE or FALSE.
}
\description{
Check if Object is an S3 Store Component
}
