% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{print.datom_conn}
\alias{print.datom_conn}
\title{Print a datom Connection}
\usage{
\method{print}{datom_conn}(x, ...)
}
\arguments{
\item{x}{A \code{datom_conn} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Displays a clean summary without exposing credentials or the S3 client.
}
