% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{print.datom_store}
\alias{print.datom_store}
\title{Print a datom Store}
\usage{
\method{print}{datom_store}(x, ...)
}
\arguments{
\item{x}{A \code{datom_store} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Displays store configuration with masked secrets.
}
