% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.datom_summary}
\alias{print.datom_summary}
\title{Print a datom_summary}
\usage{
\method{print}{datom_summary}(x, ...)
}
\arguments{
\item{x}{A \code{datom_summary} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Print a datom_summary
}
